+++ plugins/external_links/external_links.php
@@ -1,25 +1,28 @@
 <?php
 /**
- * External Links v1.2.2
+ * External Links v1.4.2
  *
  * This plugin adds small icons to external and mailto links, informing
  * users the link will take them to a new site or open their email client.
  *
- * Licensed under MIT, see LICENSE.
+ * Dual licensed under the MIT or GPL Version 3 licenses, see LICENSE.
+ * http://benjamin-regler.de/license/
  *
  * @package     External Links
- * @version     1.2.2
+ * @version     1.4.2
  * @link        <https://github.com/sommerregen/grav-plugin-external-links>
  * @author      Benjamin Regler <sommerregen@benjamin-regler.de>
  * @copyright   2015, Benjamin Regler
- * @license     <http://opensource.org/licenses/MIT>            MIT
+ * @license     <http://opensource.org/licenses/MIT>        MIT
+ * @license     <http://opensource.org/licenses/GPL-3.0>    GPLv3
  */
 
 namespace Grav\Plugin;
 
-use Grav\Common\Grav;
 use Grav\Common\Plugin;
 use Grav\Common\Page\Page;
+use Grav\Common\Data\Blueprints;
+
 use RocketTheme\Toolbox\Event\Event;
 
 /**
@@ -37,7 +40,7 @@ class ExternalLinksPlugin extends Plugin
   /**
    * Instance of ExternalLinks class
    *
-   * @var object
+   * @var \Grav\Plugin\ExternalLinks
    */
   protected $external_links;
 
@@ -55,32 +58,66 @@ class ExternalLinksPlugin extends Plugin
   public static function getSubscribedEvents()
   {
     return [
-      'onPluginsInitialized' => ['onPluginsInitialized', 0],
+      'onPluginsInitialized' => ['onPluginsInitialized', 0]
     ];
   }
 
   /**
-   * Initialize configuration.
+   * Initialize configuration
    */
   public function onPluginsInitialized()
   {
-    if ($this->isAdmin()) {
-      $this->active = false;
+    if ($this->config->get('plugins.external_links.enabled')) {
+      // Set default events
+      $events = [
+        'onTwigInitialized' => ['onTwigInitialized', 0],
+        'onTwigSiteVariables' => ['onTwigSiteVariables', 0],
+        'onBuildPagesInitialized' => ['onBuildPagesInitialized', 0]
+      ];
+
+      // Set admin specific events
+      if ($this->isAdmin()) {
+        $this->active = false;
+        $events = [
+          'onBlueprintCreated' => ['onBlueprintCreated', 0]
+        ];
+      }
+
+      // Register events
+      $this->enable($events);
+    }
+  }
+
+  /**
+   * Initialize configuration when building pages.
+   */
+  public function onBuildPagesInitialized()
+  {
+    if (!$this->active) {
       return;
     }
 
-    if ($this->config->get('plugins.external_links.enabled')) {
-      // Initialize ExternalLinks class
-      require_once(__DIR__ . '/classes/ExternalLinks.php');
-      $this->external_links = new ExternalLinks();
+    // Process contents order according to weight option
+    $weight = $this->config->get('plugins.external_links.weight', 0);
 
-      // Process contents order according to weight option
-      $weight = $this->config->get('plugins.external_links.weight');
+    $this->enable([
+      'onPageContentProcessed' => ['onPageContentProcessed', $weight]
+    ]);
+  }
 
-      $this->enable([
-        'onPageContentProcessed' => ['onPageContentProcessed', $weight],
-        'onTwigSiteVariables' => ['onTwigSiteVariables', 0],
-      ]);
+  /**
+   * Extend page blueprints with ExternalLinks configuration options.
+   *
+   * @param Event $event
+   */
+  public function onBlueprintCreated(Event $event)
+  {
+    /** @var Blueprints $blueprint */
+    $blueprint = $event['blueprint'];
+    if ($blueprint->get('form.fields.tabs')) {
+      $blueprints = new Blueprints(__DIR__ . '/blueprints/');
+      $extends = $blueprints->get($this->name);
+      $blueprint->extend($extends, true);
     }
   }
 
@@ -97,15 +134,13 @@ class ExternalLinksPlugin extends Plugin
     $page = $event['page'];
 
     $config = $this->mergeConfig($page);
-    if ($config->get('process', false) && $this->compileOnce($page)) {
-      // Do nothing, if a route for a given page does not exist
-      if (!$page->route()) {
-        return;
-      }
+    $enabled = ($config->get('process') && $config->get('enabled')) ? true : false;
 
-      // Check if mode option is valid
-      $mode = $config->get('mode', 'passive');
-      if (!in_array($mode, array('active', 'passive'))) {
+    if ($enabled && $this->compileOnce($page)) {
+      // Do nothing, if a route for a given page does not exist and check if
+      // mode option is valid
+      $mode = strtolower($config->get('mode', 'passive'));
+      if (!$page->route() || !in_array($mode, array('active', 'passive'))) {
         return;
       }
 
@@ -114,9 +149,20 @@ class ExternalLinksPlugin extends Plugin
 
       // Apply external links filter and save modified page content
       $page->setRawContent(
-        $this->external_links->process($content, $config)
+        $this->externalLinksFilter($content, $config->toArray(), $page)
       );
     }
+  }
+
+  /**
+   * Initialize Twig configuration and filters.
+   */
+  public function onTwigInitialized()
+  {
+    // Expose function
+    $this->grav['twig']->twig()->addFilter(
+      new \Twig_SimpleFilter('external_links', [$this, 'externalLinksFilter'], ['is_safe' => ['html']])
+    );
   }
 
   /**
@@ -124,9 +170,27 @@ class ExternalLinksPlugin extends Plugin
    */
   public function onTwigSiteVariables()
   {
-    if ( $this->config->get('plugins.external_links.built_in_css') ) {
+    if ($this->config->get('plugins.external_links.built_in_css')) {
       $this->grav['assets']->add('plugin://external_links/assets/css/external_links.css');
     }
+  }
+
+  /**
+   * Filter to parse external links.
+   *
+   * @param  string $content The content to be filtered.
+   * @param  array  $options Array of options for the External links filter.
+   *
+   * @return string          The filtered content.
+   */
+  public function externalLinksFilter($content, $params = [])
+  {
+    // Get custom user configuration
+    $page = func_num_args() > 2 ? func_get_arg(2) : $this->grav['page'];
+    $config = $this->mergeConfig($page, true, $params);
+
+    // Render
+    return $this->init()->render($content, $config, $page);
   }
 
   /** -------------------------------
@@ -153,5 +217,21 @@ class ExternalLinksPlugin extends Plugin
     }
 
     return false;
+  }
+
+  /**
+   * Initialize plugin and all dependencies.
+   *
+   * @return \Grav\Plugin\ExternalLinks   Returns ExternalLinks instance.
+   */
+  protected function init()
+  {
+    if (!$this->external_links) {
+      // Initialize ExternalLinks instance
+      require_once(__DIR__ . '/classes/ExternalLinks.php');
+      $this->external_links = new ExternalLinks();
+    }
+
+    return $this->external_links;
   }
 }