+++ plugins/problems/problems.php
@@ -17,7 +17,7 @@ class ProblemsPlugin extends Plugin
     public static function getSubscribedEvents()
     {
         return [
-            'onPluginsInitialized' => ['onPluginsInitialized', 0],
+            'onPluginsInitialized' => ['onPluginsInitialized', 100001],
             'onFatalException' => ['onFatalException', 0]
         ];
     }
@@ -80,10 +80,10 @@ class ProblemsPlugin extends Plugin
 
         $problems = '';
         foreach ($this->results as $key => $result) {
-            if ($key == 'files') {
-                foreach ($result as $filename => $file_result) {
-                    foreach ($file_result as $status => $text) {
-                        $problems .= $this->getListRow($status, '<b>' . $filename . '</b> ' . $text);
+            if ($key == 'files' || $key == 'apache' || $key == 'execute') {
+                foreach ($result as $key_text => $value_text) {
+                    foreach ($value_text as $status => $text) {
+                        $problems .= $this->getListRow($status, '<b>' . $key_text . '</b> ' . $text);
                     }
                 }
             } else {
@@ -121,7 +121,7 @@ class ProblemsPlugin extends Plugin
 
     protected function problemChecker()
     {
-        $min_php_version = '5.4.0';
+        $min_php_version = defined('GRAV_PHP_MIN') ? GRAV_PHP_MIN : '5.4.0';
         $problems_found = false;
 
         $essential_files = [
@@ -148,6 +148,31 @@ class ProblemsPlugin extends Plugin
             }
         }
 
+        // Perform some Apache checks
+        if (strpos(php_sapi_name(), 'apache') !== false) {
+
+            $require_apache_modules = ['mod_rewrite'];
+            $apache_modules = apache_get_modules();
+
+            $apache_status = [];
+
+            foreach ($require_apache_modules as $module) {
+                if (in_array($module, $apache_modules)) {
+                    $apache_module_adjective = ' Apache module is enabled';
+                    $apache_module_status = 'success';
+                } else {
+                    $problems_found = true;
+                    $apache_module_adjective = ' Apache module is not installed or enabled';
+                    $apache_module_status = 'error';
+                }
+                $apache_status[$module] = [$apache_module_status => $apache_module_adjective];
+            }
+
+            if (sizeof($apache_status) > 0) {
+                $this->results['apache'] = $apache_status;
+            }
+        }
+
         // Check PHP version
         if (version_compare(phpversion(), $min_php_version, '<')) {
             $problems_found = true;
@@ -158,7 +183,7 @@ class ProblemsPlugin extends Plugin
             $php_version_adjective = 'greater';
             $php_version_status = 'success';
         }
-        $this->results['php'] = [$php_version_status => 'Your PHP version (' . phpversion() . ') is '. $php_version_adjective . ' than the minimum required: <b>' . $min_php_version . '</b>'];
+        $this->results['php'] = [$php_version_status => 'Your PHP version (' . phpversion() . ') is '. $php_version_adjective . ' than the minimum required: <b>' . $min_php_version . '</b>  - <a href="http://getgrav.org/blog/changing-php-requirements-to-5.5">Additional Information</a>'];
 
         // Check for GD library
         if (defined('GD_VERSION') && function_exists('gd_info')) {
@@ -181,6 +206,28 @@ class ProblemsPlugin extends Plugin
             $curl_status = 'error';
         }
         $this->results['curl'] = [$curl_status => 'PHP Curl (Data Transfer Library) is '. $curl_adjective . 'installed'];
+
+        // Check for PHP Open SSL library
+        if (extension_loaded('openssl') && defined('OPENSSL_VERSION_TEXT')) {
+            $ssl_adjective = '';
+            $ssl_status = 'success';
+        } else {
+            $problems_found = true;
+            $ssl_adjective = 'not ';
+            $ssl_status = 'error';
+        }
+        $this->results['ssl'] = [$ssl_status => 'PHP OpenSSL (Secure Sockets Library) is '. $ssl_adjective . 'installed'];
+
+        // Check for PHP MbString library
+        if (extension_loaded('mbstring')) {
+            $mbstring_adjective = '';
+            $mbstring_status = 'success';
+        } else {
+            $problems_found = true;
+            $mbstring_adjective = 'not ';
+            $mbstring_status = 'error';
+        }
+        $this->results['mbstring'] = [$mbstring_status => 'PHP Mbstring (Multibyte String Library) is '. $mbstring_adjective . 'installed'];
 
         // Check for essential files & perms
         $file_problems = [];